requirements.txt
requests==2.32.3
langchain==0.2.3
langchain-openai==0.1.8
python-dotenv==1.0.1
langchain-community==0.2.4